# api/openapi/competency-constructor-v1.yaml
openapi: 3.0.3
info:
  title: Constructor competence API
  description: The service is designed to create and fill a bank of competencies
  version: 0.0.1
servers:
  - url: "https://{hostname}/competency-constructor/api/v1"
paths:
  /levels:
    get:
      tags:
        - levels
      operationId: getLevels
      summary: Returns level of the educational program.
      responses:
        200:
          description: Found previously created level educational.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LevelResponse"
    post:
      tags:
        - levels
      operationId: createLevel
      summary: Create the level of the educational program
      requestBody:
        description: Level to create.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLevelRequest"
      responses:
        201:
          description: Level is created.
          headers:
            Location:
              description: Created level URI.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /levels/{levelId}:
    get:
      tags:
        - levels
      operationId: getSpecificLevel
      summary: Returns level with such ID.
      parameters:
        - in: path
          name: levelId
          schema:
            type: string
            format: uuid
          required: true
          description: Level ID to return
      responses:
        200:
          description: Found previously created level.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecificLevelResponse"
        404:
          description: Level with such ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    Error:
      type: object
      required:
        - slug
        - details
      properties:
        slug:
          $ref: "#/components/schemas/ErrorSlug"
        details:
          type: string
    ErrorSlug:
      type: string
      enum:
        - level-not-found
        - ugsn-not-found
        - specialty-not-found
        - program-not-found
        - competency-not-found
        - indicator-not-found
        - invalid-json
        - bad-request
        - unexpected-error
        - empty-bearer-token
        - unable-to-verify-jwt
        - unauthorized-user
        - invalid-level-parameters
        - invalid-ugsn-parameters
        - invalid-specialties-parameters
        - invalid-program-parameters
        - invalid-competency-parameters
        - invalid-indicator-parameters
...
    CreateLevelRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
    LevelResponse:
      type: object
      required:
        - id
        - title
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
    SpecificLevelResponse:
      type: object
      required:
        - id
        - title
        - ugsn
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        ugsn:
          type: array
          items:
            $ref: "#/components/schemas/SpecificUgsnResponse"
    ProgramResponse:
      type: object
      required:
        - id
        - code
        - title
        - specialtyId
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          pattern: '^(0[1-9]{1}|[1-9]{1}[0-9]{1})\.(0[1-9]{1}|[1-9]{1}[0-9]{1})\.(0[1-9]{1}|[1-9]{1}[0-9]{1})-(0[1-9]{1}|[1-9]{1}[0-9]{1})$'
        title:
          type: string
        specialtyId:
          type: string
          format: uuid
    SpecificSpecialtyResponse:
      type: object
      required:
        - id
        - code
        - title
        - program
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          pattern: '^(0[1-9]{1}|[1-9]{1}[0-9]{1})\.(0[1-9]{1}|[1-9]{1}[0-9]{1})\.(0[1-9]{1}|[1-9]{1}[0-9]{1})$'
        title:
          type: string
        program:
          type: array
          items:
            $ref: "#/components/schemas/ProgramResponse"
    SpecificUgsnResponse:
      type: object
      required:
        - id
        - code
        - title
        - specialty
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          pattern: '^(0[1-9]{1}|[1-9]{1}[0-9]{1})\.00\.00$'
        title:
          type: string
        specialty:
          type: array
          items:
            $ref: "#/components/schemas/SpecificSpecialtyResponse"
...