# internal/core/adapter/driver/rest/v1/openapi_type.gen.go
// Package v1 provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.12.3 DO NOT EDIT.
package v1

import (
	openapi_types "github.com/deepmap/oapi-codegen/pkg/types"
)

// Defines values for ErrorSlug.
const (
	BadRequest                   ErrorSlug = "bad-request"
	CompetencyNotFound           ErrorSlug = "competency-not-found"
	EmptyBearerToken             ErrorSlug = "empty-bearer-token"
	IndicatorNotFound            ErrorSlug = "indicator-not-found"
	InvalidCompetencyParameters  ErrorSlug = "invalid-competency-parameters"
	InvalidIndicatorParameters   ErrorSlug = "invalid-indicator-parameters"
	InvalidJson                  ErrorSlug = "invalid-json"
	InvalidLevelParameters       ErrorSlug = "invalid-level-parameters"
	InvalidProgramParameters     ErrorSlug = "invalid-program-parameters"
	InvalidSpecialtiesParameters ErrorSlug = "invalid-specialties-parameters"
	InvalidUgsnParameters        ErrorSlug = "invalid-ugsn-parameters"
	LevelNotFound                ErrorSlug = "level-not-found"
	ProgramNotFound              ErrorSlug = "program-not-found"
	SpecialtyNotFound            ErrorSlug = "specialty-not-found"
	UgsnNotFound                 ErrorSlug = "ugsn-not-found"
	UnableToVerifyJwt            ErrorSlug = "unable-to-verify-jwt"
	UnauthorizedUser             ErrorSlug = "unauthorized-user"
	UnexpectedError              ErrorSlug = "unexpected-error"
)

// Defines values for Type.
const (
	ОПК Type = "ОПК"
	ПК  Type = "ПК"
	УК  Type = "УК"
)

// CompetencyFilterRequest defines model for CompetencyFilterRequest.
type CompetencyFilterRequest struct {
	LevelId     *openapi_types.UUID `json:"levelId,omitempty"`
	ProgramId   *openapi_types.UUID `json:"programId,omitempty"`
	SpecialtyId *openapi_types.UUID `json:"specialtyId,omitempty"`
	UgsnId      *openapi_types.UUID `json:"ugsnId,omitempty"`
}

// CompetencyResponse defines model for CompetencyResponse.
type CompetencyResponse struct {
	Category    string              `json:"category"`
	Code        string              `json:"code"`
	Id          openapi_types.UUID  `json:"id"`
	LevelId     *openapi_types.UUID `json:"levelId,omitempty"`
	ProgramId   *openapi_types.UUID `json:"programId,omitempty"`
	SpecialtyId *openapi_types.UUID `json:"specialtyId,omitempty"`
	Title       string              `json:"title"`
	Type        Type                `json:"type"`
	UgsnId      *openapi_types.UUID `json:"ugsnId,omitempty"`
}

// CreateCompetencyRequest defines model for CreateCompetencyRequest.
type CreateCompetencyRequest struct {
	Category    string              `json:"category"`
	Code        string              `json:"code"`
	LevelId     *openapi_types.UUID `json:"levelId,omitempty"`
	ProgramId   *openapi_types.UUID `json:"programId,omitempty"`
	SpecialtyId *openapi_types.UUID `json:"specialtyId,omitempty"`
	Title       string              `json:"title"`
	Type        Type                `json:"type"`
	UgsnId      *openapi_types.UUID `json:"ugsnId,omitempty"`
}

// CreateIndicatorRequest defines model for CreateIndicatorRequest.
type CreateIndicatorRequest struct {
	Code  string `json:"code"`
	Title string `json:"title"`
}

// CreateLevelRequest defines model for CreateLevelRequest.
type CreateLevelRequest struct {
	Title string `json:"title"`
}

// CreateProgramRequest defines model for CreateProgramRequest.
type CreateProgramRequest struct {
	Code  string `json:"code"`
	Title string `json:"title"`
}

// CreateSpecialtyRequest defines model for CreateSpecialtyRequest.
type CreateSpecialtyRequest struct {
	Code  string `json:"code"`
	Title string `json:"title"`
}

// CreateUgsnRequest defines model for CreateUgsnRequest.
type CreateUgsnRequest struct {
	Code  string `json:"code"`
	Title string `json:"title"`
}

// Error defines model for Error.
type Error struct {
	Details string    `json:"details"`
	Slug    ErrorSlug `json:"slug"`
}

// ErrorSlug defines model for ErrorSlug.
type ErrorSlug string

// LevelResponse defines model for LevelResponse.
type LevelResponse struct {
	Id    openapi_types.UUID `json:"id"`
	Title string             `json:"title"`
}

// ProgramResponse defines model for ProgramResponse.
type ProgramResponse struct {
	Code        string             `json:"code"`
	Id          openapi_types.UUID `json:"id"`
	SpecialtyId openapi_types.UUID `json:"specialtyId"`
	Title       string             `json:"title"`
}

// SpecialtyResponse defines model for SpecialtyResponse.
type SpecialtyResponse struct {
	Code   string             `json:"code"`
	Id     openapi_types.UUID `json:"id"`
	Title  string             `json:"title"`
	UgsnId openapi_types.UUID `json:"ugsnId"`
}

// SpecificLevelResponse defines model for SpecificLevelResponse.
type SpecificLevelResponse struct {
	Id    openapi_types.UUID     `json:"id"`
	Title string                 `json:"title"`
	Ugsn  []SpecificUgsnResponse `json:"ugsn"`
}

// SpecificSpecialtyResponse defines model for SpecificSpecialtyResponse.
type SpecificSpecialtyResponse struct {
	Code    string             `json:"code"`
	Id      openapi_types.UUID `json:"id"`
	Program []ProgramResponse  `json:"program"`
	Title   string             `json:"title"`
}

// SpecificUgsnResponse defines model for SpecificUgsnResponse.
type SpecificUgsnResponse struct {
	Code      string                      `json:"code"`
	Id        openapi_types.UUID          `json:"id"`
	Specialty []SpecificSpecialtyResponse `json:"specialty"`
	Title     string                      `json:"title"`
}

// Type defines model for Type.
type Type string

// UgsnResponse defines model for UgsnResponse.
type UgsnResponse struct {
	Code    string             `json:"code"`
	Id      openapi_types.UUID `json:"id"`
	LevelId openapi_types.UUID `json:"levelId"`
	Title   string             `json:"title"`
}

// FilterCompetencyJSONRequestBody defines body for FilterCompetency for application/json ContentType.
type FilterCompetencyJSONRequestBody = CompetencyFilterRequest

// CreateCompetencyJSONRequestBody defines body for CreateCompetency for application/json ContentType.
type CreateCompetencyJSONRequestBody = CreateCompetencyRequest

// CreateIndicatorJSONRequestBody defines body for CreateIndicator for application/json ContentType.
type CreateIndicatorJSONRequestBody = CreateIndicatorRequest

// CreateLevelJSONRequestBody defines body for CreateLevel for application/json ContentType.
type CreateLevelJSONRequestBody = CreateLevelRequest

// AddUgsnJSONRequestBody defines body for AddUgsn for application/json ContentType.
type AddUgsnJSONRequestBody = CreateUgsnRequest

// AddProgramsJSONRequestBody defines body for AddPrograms for application/json ContentType.
type AddProgramsJSONRequestBody = CreateProgramRequest

// AddSpecialtiesJSONRequestBody defines body for AddSpecialties for application/json ContentType.
type AddSpecialtiesJSONRequestBody = CreateSpecialtyRequest